backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pdf2image
pytesseract
python-dotenv
google-generativeai
google-generativeai

